+++ net/mptcp/protocol.h
@@ -263,7 +263,7 @@ struct mptcp_data_frag {
 	struct list_head list;
 	u64 data_seq;
 	u16 data_len;
-	u16 offset;
+	u32 offset;		/* u16 wraps on 256KB-page kernels */
 	u8 overhead;
 	u8 eor;			/* currently using 1 bit */
 	u16 already_sent;
